pyproject.toml
[project]
name = "planning_playground"
version = "0.0.1"
dependencies = [
    "matplotlib",
    "numpy",
    "pgraph-python @ git+https://github.com/petercorke/pgraph-python.git"
]
